+++ components/serdes/src/br/dev/yuhri/serdes/formats/json.clj
@@ -1,0 +1,19 @@
+(ns br.dev.yuhri.serdes.formats.json
+  (:require [muuntaja.core :as mtj]
+            [br.dev.yuhri.serdes.content-negotiation :as content-negotiation]))
+
+(def ^:private default-muuntaja content-negotiation/muuntaja)
+
+(defn json->clj
+  ([v]
+   (mtj/decode
+     default-muuntaja
+     "application/json" v)))
+
+(defn clj->json-stream [v]
+  (mtj/encode
+    default-muuntaja
+    "application/json" v))
+
+(defn clj->json [v]
+  (slurp (clj->json-stream v)))